Summary: <REQUIRED Add a brief Summary here> Comments: <OPTIONAL Add additional comments as needed>

+++ iaas/iaas-slamanager/planning-optimization/src/main/java/eu/ascetic/iaas/slamanager/poc/manager/negotiation/NegotiationManagerImpl.java
@@ -389,7 +389,7 @@ public class NegotiationManagerImpl implements NegotiationManager {
 							SimpleDomainExpr sde = (SimpleDomainExpr)tce.getDomain();
 
 							CONST originalValue = (CONST) sde.getValue();
-							Integer mathValue = Integer.parseInt(originalValue.getValue());
+							Double mathValue = Double.parseDouble(originalValue.getValue());
 
 							mathValue = mathValue / ovfParser.getVirtualSystems().length;
